backend/requirements.txt
openai
Flask
azure-storage-blob
azure.identity
python-dotenv
webuiapi
flask-session
werkzeug